+++ Mecanillama.API/Appointments/Resources/AppointmentResource.cs
@@ -2,9 +2,9 @@
 
 public class AppointmentResource
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Date { get; set; }
     public string Time { get; set; }
-    public long CustomerId { get; set; }
-    public long MechanicId { get; set; }
+    public int CustomerId { get; set; }
+    public int MechanicId { get; set; }
 }
